Have common only reference within common
Part of #1337
