+++ src/common/CircularList.ts
@@ -3,8 +3,8 @@
  * @license MIT
  */
 
-import { EventEmitter } from '../EventEmitter';
-import { ICircularList } from '../Types';
+import { EventEmitter } from './EventEmitter';
+import { ICircularList } from './Types';
 
 /**
  * Represents a circular list; a list with a maximum size that wraps around when push is called,
